move MethodTable flags and validation to a separate classes
